pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alphatales-refactor-agent"
version = "2.0.0"
description = "AI-powered code refactoring system using Claude Agent SDK"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "AlphaTales Team" }
]
keywords = ["refactoring", "claude", "ai", "code-quality", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    # Core - Claude Agent SDK
    "claude-agent-sdk>=0.1.14",
    "anthropic>=0.75.0",

    # CLI & Console
    "typer>=0.12.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "textual>=0.89.0",
    "textual-autocomplete>=4.0.0",

    # Configuration
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Logging
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.8.0",
]
analysis = [
    "tree-sitter>=0.21.0",
    "pygments>=2.17.0",
    "radon>=6.0.1",
]

[project.scripts]
refactor-agent = "refactor_agent.cli:main"
alphatales-refactor = "refactor_agent.cli:main"

[project.urls]
Homepage = "https://github.com/alphatales/refactor-agent"
Documentation = "https://github.com/alphatales/refactor-agent#readme"
Issues = "https://github.com/alphatales/refactor-agent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/refactor_agent"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/rules",
    "/.claude",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["refactor_agent"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
